[REF] refator class andteacher deletion

+++ src/controller/class.controller.ts
@@ -177,23 +177,43 @@ router.patch(
   }
 );
 
-//   delete class
-
+// delete class
 router.delete(
   "/class/:id",
   async (req: Request, res: Response, next: NextFunction) => {
     try {
       const { id } = req.params;
-      const grade = await prisma.class.delete({
+
+      const classToDelete = await prisma.class.findUnique({
+        where: {
+          id: Number(id),
+        },
+        include: {
+          Teacher: true,
+        },
+      });
+
+      if (!classToDelete) {
+        return res.status(404).json({ message: "Class not found" });
+      }
+
+      const { Teacher } = classToDelete;
+
+      if (Teacher) {
+        return res
+          .status(400)
+          .json({ message: "Cannot delete a class that has a teacher" });
+      }
+
+      await prisma.class.delete({
         where: {
           id: Number(id),
         },
       });
 
-      res.status(204).json(grade);
+      res.status(204).end();
     } catch (error) {
       next(error);
-      return res.status(404).json({ message: "Class not found" });
     }
   }
 );
